Add UI Login, OTP, Register, Onboarding
