+++ rn/colors.ts
@@ -1,0 +1,10 @@
+import { Colors } from 'react-native-ui-lib';
+
+Colors.loadColors({
+  primary: '#717658',
+  secondary: 'white',
+  black: 'black',
+  gray: '#AAAAAA',
+});
+
+export default Colors;
